requirements.txt
Flask==1.1.2
Flask-Testing==0.8.0
Flask-WTF==0.14.3
pytest==6.2.1
requests==2.25.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
WTForms==2.3.3
gunicorn==20.0.4
pyyaml==5.4.1
coloredlogs==15.0